+++ openeo/__init__.py
@@ -8,6 +8,10 @@ This client allows users to communicate with OpenEO backends, in a way that feel
     :caption: Basic example
     :name: basic_example
 
+.. image:: ../docs/images/basic_example.png
+  :width: 500
+  :alt: script result
+
 
 In this example, a 'datacube' is created using `load_collection`. Datacubes can be transformed by applying operations directly.